Sm/command-renaming (#673)
* feat: finish sf-ification

* chore: github cleanup
