+++ .git2gus/config.json
@@ -1,6 +1,6 @@
 {
   "productTag": "a1aB00000004Bx8IAE",
-  "defaultBuild": "offcore.tooling.56",
+  "defaultBuild": "offcore.tooling.59",
   "issueTypeLabels": {
     "feature": "USER STORY",
     "regression": "BUG P1",
