+++ package.json
@@ -32,7 +32,7 @@
     "eslint-plugin-header": "^3.1.1",
     "eslint-plugin-import": "2.28.1",
     "eslint-plugin-jsdoc": "^46.8.2",
-    "eslint-plugin-sf-plugin": "^1.16.9",
+    "eslint-plugin-sf-plugin": "^1.16.12",
     "husky": "^7.0.4",
     "mocha": "^9.1.3",
     "nyc": "^15.1.0",
@@ -78,14 +78,16 @@
       "@salesforce/plugin-command-reference"
     ],
     "topics": {
-      "limits": {
-        "description": "Display an org’s limits.",
+      "org": {
+        "external": true,
         "subtopics": {
-          "api": {
-            "description": "Display an org’s API limits."
+          "list": {
+            "external": true
           },
-          "recordcounts": {
-            "description": "Display record counts for an org."
+          "subtopics": {
+            "sobject": {
+              "external": true
+            }
           }
         }
       }
