+++ src/commands/org/list/limits.ts
@@ -33,7 +33,7 @@ interface Result {
 export type ApiLimits = ApiLimit[];
 
 export class LimitsApiDisplayCommand extends SfCommand<ApiLimits> {
-  public static readonly aliases = ['force:limits:api:display', 'org:list:limits'];
+  public static readonly aliases = ['force:limits:api:display', 'limits:api:display'];
   public static readonly summary = messages.getMessage('summary');
   public static readonly description = messages.getMessage('description');
   public static readonly examples = messages.getMessages('examples');
