+++ src/commands/org/list/sobject/record-counts.ts
@@ -30,7 +30,7 @@ interface Result {
 }
 
 export class LimitsRecordCountsDisplayCommand extends SfCommand<RecordCounts> {
-  public static readonly aliases = ['force:limits:recordcounts:display', 'org:list:sobject:record-counts'];
+  public static readonly aliases = ['force:limits:recordcounts:display', 'limits:recordcounts:display'];
   public static readonly summary = messages.getMessage('summary');
   public static readonly description = messages.getMessage('description');
   public static readonly examples = messages.getMessages('examples');
@@ -38,7 +38,7 @@ export class LimitsRecordCountsDisplayCommand extends SfCommand<RecordCounts> {
   public static readonly flags = {
     sobject: arrayWithDeprecation({
       char: 's',
-      summary: messages.getMessage('sobjectFlagDescription'),
+      summary: messages.getMessage('flags.sobject.summary'),
       aliases: ['sobjecttype'],
       default: [],
     }),
